# Blog Project（Frontend）/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BinSpace API
  version: 2.0.0-draft.1
  description: |
    BinSpace 当前 42 个接口基线与下一阶段功能 2～8 契约。
    认证使用 Spring Security、Bearer JWT 和 Redis Token Session。
    新增接口在后端实现前仅代表约定，不代表当前环境已经可调用。
servers:
  - url: http://localhost:8080
    description: Local Spring Boot
...
paths:
...
  /api/posts:
    get:
      summary: 获取首页混合 Feed
      description: 只返回公开的 PUBLISHED TECH/MOMENT；无请求体。
      tags: [Posts]
      operationId: listPublicPosts
      parameters:
        - { $ref: '#/components/parameters/PostTypeQuery' }
        - { $ref: '#/components/parameters/PageQuery' }
        - { $ref: '#/components/parameters/PageSizeQuery' }
      responses:
        '200':
          description: 分页 Post 摘要
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
      security: []
  /api/posts/tech:
    get:
      summary: 获取公开 TECH 列表
      description: 支持 Category/Tag slug 筛选，只返回公开文章；无请求体。
      tags: [Tech]
      operationId: listPublicTechPosts
      parameters:
        - { $ref: '#/components/parameters/CategorySlugQuery' }
        - { $ref: '#/components/parameters/TagSlugQuery' }
        - { $ref: '#/components/parameters/PageQuery' }
        - { $ref: '#/components/parameters/PageSizeQuery' }
      responses:
        '200':
          description: 分页 TECH 摘要
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
      security: []
  /api/posts/tech/{slug}:
    get:
      summary: 获取公开 TECH 详情
      description: 通过稳定 slug 获取正文、关联元信息和相关文章；无请求体。
      tags: [Tech]
      operationId: getPublicTechPost
      parameters:
        - { $ref: '#/components/parameters/SlugPath' }
      responses:
        '200':
          description: TECH 详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TechDetailResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
      security: []
  /api/posts/moments:
    get:
      summary: 获取公开 MOMENT 列表
      description: 前端按 publishedAt 分月展示；无请求体。
      tags: [Moments]
      operationId: listPublicMoments
      parameters:
        - { $ref: '#/components/parameters/YearQuery' }
        - { $ref: '#/components/parameters/MonthQuery' }
        - { $ref: '#/components/parameters/PageQuery' }
        - { $ref: '#/components/parameters/PageSizeQuery' }
      responses:
        '200':
          description: 分页 MOMENT 摘要
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
      security: []
  /api/posts/moments/{id}:
    get:
      summary: 获取公开 MOMENT 详情
      description: MOMENT 不返回标题、摘要、分类和阅读时间；无请求体。
      tags: [Moments]
      operationId: getPublicMoment
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: MOMENT 详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MomentDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
      security: []
...
  /api/owner/posts:
    get:
      summary: 获取 Owner 内容列表
      description: 可读取 DRAFT/PUBLISHED，供刷新后恢复草稿；无请求体。
      tags: [Owner Posts]
      operationId: listOwnerPosts
      parameters:
        - { $ref: '#/components/parameters/PostTypeQuery' }
        - { $ref: '#/components/parameters/PostStatusQuery' }
        - { $ref: '#/components/parameters/PageQuery' }
        - { $ref: '#/components/parameters/PageSizeQuery' }
      responses:
        '200':
          description: Owner 分页内容，包含 status/version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/posts/tech:
    post:
      summary: 创建 TECH
      description: 服务端生成 id、slug、readingTime、时间和互动计数；正文图片 URL 保存在 content。
      tags: [Owner Posts]
      operationId: createTechPost
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTechPostDTO' }
      responses:
        '201':
          description: 创建后的 TECH
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TechDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/posts/tech/{id}:
    get:
      summary: Owner 按 ID 获取 TECH
      description: 可读取 DRAFT/PUBLISHED，返回 status/version；无请求体。
      tags: [Owner Posts]
      operationId: getOwnerTechPost
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: Owner TECH 编辑详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TechDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
    put:
      summary: 更新 TECH
      description: 完整替换可编辑字段；标题变化不修改 slug，必须携带 version。
      tags: [Owner Posts]
      operationId: updateTechPost
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTechPostDTO' }
      responses:
        '200':
          description: 更新后的 TECH
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TechDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
    delete:
      summary: 删除 TECH
      description: 使用 id + version 逻辑删除；不使用 ETag/If-Match；无请求体。
      tags: [Owner Posts]
      operationId: deleteTechPost
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
        - { $ref: '#/components/parameters/VersionQuery' }
      responses:
        '204': { description: 删除成功，无响应体 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/posts/moments:
    post:
      summary: 创建 MOMENT
      description: MOMENT 只接收 content、images、status，不接收 TECH 专属字段。
      tags: [Owner Posts]
      operationId: createMoment
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMomentDTO' }
      responses:
        '201':
          description: 创建后的 MOMENT
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MomentDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/posts/moments/{id}:
    get:
      summary: Owner 按 ID 获取 MOMENT
      description: 可读取 DRAFT/PUBLISHED，返回 status/version；无请求体。
      tags: [Owner Posts]
      operationId: getOwnerMoment
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: Owner MOMENT 编辑详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MomentDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
    put:
      summary: 更新 MOMENT
      description: 完整替换 content/images/status，必须携带 version。
      tags: [Owner Posts]
      operationId: updateMoment
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMomentDTO' }
      responses:
        '200':
          description: 更新后的 MOMENT
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MomentDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
    delete:
      summary: 删除 MOMENT
      description: 使用 id + version 逻辑删除；不使用 ETag/If-Match；无请求体。
      tags: [Owner Posts]
      operationId: deleteMoment
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
        - { $ref: '#/components/parameters/VersionQuery' }
      responses:
        '204': { description: 删除成功，无响应体 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "登录/注册接口返回的 JWT Access Token。Visitor 与 Owner 受保护请求均使用 Authorization: Bearer <token>。"
  parameters:
    PageQuery:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
    IdPath:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/BigIntId' }
...
    SlugPath:
      name: slug
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 180 }
      example: redis-lua-stock-control
    VersionQuery:
      name: version
      in: query
      required: true
      schema: { type: integer, minimum: 0 }
      example: 3
    PostTypeQuery:
      name: type
      in: query
      required: false
      schema: { type: string, enum: [ALL, TECH, MOMENT], default: ALL }
    PostStatusQuery:
      name: status
      in: query
      required: false
      schema: { type: string, enum: [ALL, DRAFT, SCHEDULED, PUBLISHED], default: ALL }
    CategorySlugQuery:
      name: category
      in: query
      required: false
      schema: { type: string }
      example: redis
    TagSlugQuery:
      name: tag
      in: query
      required: false
      schema: { type: string }
      example: lua
    YearQuery:
      name: year
      in: query
      required: false
      schema: { type: integer, minimum: 1970, maximum: 9999 }
      example: 2026
    MonthQuery:
      name: month
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 12 }
      example: 9
...
  responses:
    BadRequest:
      description: 请求参数或 Bean Validation 失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: BAD_REQUEST, message: 请求参数不正确, data: { title: 标题不能为空 }, traceId: 01JEXAMPLE }
    Unauthorized:
      description: 未携带 Token、JWT 无效/过期，或 Redis Token Session 已失效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: AUTH_REQUIRED, message: 请先登录, data: null, traceId: 01JEXAMPLE }
...
    Forbidden:
      description: 当前 JWT 身份缺少接口要求的角色或权限
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: FORBIDDEN, message: 无权执行此操作, data: null, traceId: 01JEXAMPLE }
    NotFound:
      description: 资源不存在、已删除或不可见
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: RESOURCE_NOT_FOUND, message: 资源不存在, data: null, traceId: 01JEXAMPLE }
    VersionConflict:
      description: 乐观锁版本冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: VERSION_CONFLICT, message: 数据已被修改, data: null, traceId: 01JEXAMPLE }
    ResourceConflict:
      description: 唯一键、重复回复或资源状态冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: RESOURCE_CONFLICT, message: 资源状态冲突, data: null, traceId: 01JEXAMPLE }
    Conflict:
      description: 乐观锁、唯一键或资源状态冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            version:
              value: { code: VERSION_CONFLICT, message: 数据已被修改, data: null, traceId: 01JEXAMPLE }
            resource:
              value: { code: RESOURCE_CONFLICT, message: 资源状态冲突, data: null, traceId: 01JEXAMPLE }
    InternalError:
      description: 未处理的服务端错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: INTERNAL_ERROR, message: 服务暂时不可用, data: null, traceId: 01JEXAMPLE }
...
  schemas:
    NullOnly:
      type: object
      nullable: true
      enum: [null]
      description: OpenAPI 3.0 中供 anyOf 复用的 null 分支。
    BigIntId:
      type: string
      pattern: '^[0-9]+$'
      description: MySQL BIGINT 的十进制字符串，禁止在前端转换为 Number。
      example: '1909508401234567168'
    ApiSuccessBase:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: string, enum: [OK], example: OK }
        message: { type: string, example: success }
        traceId: { type: string, example: 01JEXAMPLE }
    ErrorResponse:
      type: object
      required: [code, message, data, traceId]
      properties:
        code:
          type: string
          enum: [BAD_REQUEST, AUTH_REQUIRED, AUTH_FAILED, FORBIDDEN, RESOURCE_NOT_FOUND, VERSION_CONFLICT, RESOURCE_CONFLICT, PAYLOAD_TOO_LARGE, UNSUPPORTED_MEDIA_TYPE, RATE_LIMITED, UPSTREAM_ERROR, INTERNAL_ERROR]
        message: { type: string }
        data:
          type: object
          nullable: true
          additionalProperties: true
        traceId: { type: string }
...
    PostAuthorVO:
      type: object
      required: [userId, name, avatar]
      properties:
        userId: { $ref: '#/components/schemas/BigIntId' }
        name: { type: string, example: 玢 }
        avatar: { type: string, nullable: true, example: /media/avatar.jpg }
    MediaVO:
      type: object
      required: [id, usageType, mediaType, src, poster, alt, width, height, sortOrder]
      properties:
        id: { $ref: '#/components/schemas/BigIntId' }
        usageType: { type: string, enum: [COVER, CONTENT] }
        mediaType: { type: string, enum: [IMAGE, VIDEO] }
        src: { type: string, example: /media/posts/redis-cover.jpg }
        poster: { type: string, nullable: true }
        alt: { type: string, nullable: true, example: Redis 库存流程示意图 }
        width: { type: integer, nullable: true, minimum: 1, example: 1600 }
        height: { type: integer, nullable: true, minimum: 1, example: 900 }
        sortOrder: { type: integer, minimum: 0, example: 0 }
    MediaInputDTO:
      type: object
      required: [src, mediaType, poster, alt, width, height, sortOrder]
      properties:
        src: { type: string, minLength: 1, maxLength: 1024 }
        mediaType: { type: string, enum: [IMAGE, VIDEO] }
        poster: { type: string, nullable: true, maxLength: 1024 }
        alt: { type: string, nullable: true, maxLength: 255 }
        width: { type: integer, nullable: true, minimum: 1 }
        height: { type: integer, nullable: true, minimum: 1 }
        sortOrder: { type: integer, minimum: 0 }
      example: { src: /media/posts/redis-cover.jpg, mediaType: IMAGE, poster: null, alt: Redis 库存流程示意图, width: 1600, height: 900, sortOrder: 0 }
    CategoryVO:
      type: object
      required: [id, name, slug, description, active]
      properties:
        id: { $ref: '#/components/schemas/BigIntId' }
        name: { type: string, example: Redis }
        slug: { type: string, example: redis }
        description: { type: string, nullable: true, example: Redis 实践记录 }
        active: { type: boolean, example: true }
    TagVO:
      type: object
      required: [id, name, slug, active]
      properties:
        id: { $ref: '#/components/schemas/BigIntId' }
        name: { type: string, example: Lua }
        slug: { type: string, example: lua }
        active: { type: boolean, example: true }
    PostSummaryVO:
      type: object
      required: [id, type, slug, title, summary, content, author, category, tags, cover, images, createdAt, updatedAt, publishedAt, readingTime, status, likeCount, likedByMe, commentCount, version]
      properties:
        id: { $ref: '#/components/schemas/BigIntId' }
        type: { type: string, enum: [TECH, MOMENT] }
        slug: { type: string, nullable: true }
        title: { type: string, nullable: true }
        summary: { type: string, nullable: true }
        content: { type: string, nullable: true }
        author: { $ref: '#/components/schemas/PostAuthorVO' }
        category:
          anyOf:
            - { $ref: '#/components/schemas/CategoryVO' }
            - { $ref: '#/components/schemas/NullOnly' }
        tags:
          type: array
          maxItems: 10
          items: { $ref: '#/components/schemas/TagVO' }
        cover:
          anyOf:
            - { $ref: '#/components/schemas/MediaVO' }
            - { $ref: '#/components/schemas/NullOnly' }
        images:
          type: array
          items: { $ref: '#/components/schemas/MediaVO' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        publishedAt: { type: string, format: date-time, nullable: true }
        readingTime: { type: integer, minimum: 1, nullable: true }
        status: { type: string, enum: [DRAFT, SCHEDULED, PUBLISHED], nullable: true }
        likeCount: { type: integer, minimum: 0 }
        likedByMe: { type: boolean, description: 未登录时固定为 false, example: false }
        commentCount: { type: integer, minimum: 0 }
        version: { type: integer, minimum: 0, nullable: true }
    TechDetailVO:
      allOf:
        - { $ref: '#/components/schemas/PostSummaryVO' }
        - type: object
          required: [contentFormat, relatedPosts]
          properties:
            contentFormat: { type: string, enum: [MARKDOWN, PLAIN_TEXT] }
            relatedPosts:
              type: array
              items: { $ref: '#/components/schemas/PostSummaryVO' }
      example:
        id: '1909508401234567168'
        type: TECH
        slug: redis-lua-stock-control
        title: Redis + Lua 如何解决库存超卖
        summary: 用原子脚本收紧库存扣减边界。
        content: "# 库存扣减\n\n正文内容。"
        contentFormat: MARKDOWN
        author: { userId: '1909508000000000001', name: 玢, avatar: null }
        category: { id: '1909508100000000001', name: Redis, slug: redis, description: null, active: true }
        tags: [{ id: '1909508200000000001', name: Lua, slug: lua, active: true }]
        cover: null
        images: []
        createdAt: '2026-09-02T10:00:00.000Z'
        updatedAt: '2026-09-02T10:20:00.000Z'
        publishedAt: '2026-09-02T10:20:00.000Z'
        readingTime: 8
        status: PUBLISHED
        likeCount: 0
        commentCount: 0
        version: 0
        relatedPosts: []
    MomentDetailVO:
      allOf:
        - { $ref: '#/components/schemas/PostSummaryVO' }
        - type: object
          required: [contentFormat]
          properties:
            contentFormat: { type: string, enum: [PLAIN_TEXT] }
      example:
        id: '1909508401234567170'
        type: MOMENT
        slug: null
        title: null
        summary: null
        content: 今天把 BinSpace 的页面整理好了。
        contentFormat: PLAIN_TEXT
        author: { userId: '1909508000000000001', name: 玢, avatar: null }
        category: null
        tags: []
        cover: null
        images: []
        createdAt: '2026-09-02T11:00:00.000Z'
        updatedAt: '2026-09-02T11:00:00.000Z'
        publishedAt: '2026-09-02T11:00:00.000Z'
        readingTime: null
        status: PUBLISHED
        likeCount: 0
        commentCount: 0
        version: 0
    CreateTechPostDTO:
      type: object
      required: [title, summary, content, contentFormat, categoryId, tagIds, cover, status]
      properties:
        title: { type: string, minLength: 1, maxLength: 180 }
        summary: { type: string, minLength: 1, maxLength: 600 }
        content: { type: string, minLength: 1 }
        contentFormat: { type: string, enum: [MARKDOWN, PLAIN_TEXT] }
        categoryId: { $ref: '#/components/schemas/BigIntId' }
        tagIds:
          type: array
          maxItems: 10
          uniqueItems: true
          items: { $ref: '#/components/schemas/BigIntId' }
        cover:
          anyOf:
            - { $ref: '#/components/schemas/MediaInputDTO' }
            - { $ref: '#/components/schemas/NullOnly' }
        status: { type: string, enum: [DRAFT, PUBLISHED] }
      example:
        title: Redis + Lua 如何解决库存超卖
        summary: 用原子脚本收紧库存扣减边界。
        content: "# 库存扣减\n\n正文内容。"
        contentFormat: MARKDOWN
        categoryId: '1909508100000000001'
        tagIds: ['1909508200000000001']
        cover: null
        status: DRAFT
    UpdateTechPostDTO:
      allOf:
        - { $ref: '#/components/schemas/CreateTechPostDTO' }
        - type: object
          required: [version]
          properties:
            version: { type: integer, minimum: 0, example: 2 }
    CreateMomentDTO:
      type: object
      required: [content, images, status]
      properties:
        content: { type: string, minLength: 1, maxLength: 2000 }
        images:
          type: array
          maxItems: 9
          items: { $ref: '#/components/schemas/MediaInputDTO' }
        status: { type: string, enum: [DRAFT, PUBLISHED] }
      example: { content: 今天把 BinSpace 的页面整理好了。, images: [], status: PUBLISHED }
    UpdateMomentDTO:
      allOf:
        - { $ref: '#/components/schemas/CreateMomentDTO' }
        - type: object
          required: [version]
          properties:
            version: { type: integer, minimum: 0, example: 1 }
    PostPageVO:
      type: object
      required: [items, page, pageSize, total, totalPages, hasMore]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/PostSummaryVO' }
        page: { type: integer, minimum: 1, example: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 50, example: 10 }
        total: { type: integer, minimum: 0, example: 16 }
        totalPages: { type: integer, minimum: 0, example: 2 }
        hasMore: { type: boolean, example: true }
    PostPageResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/PostPageVO' } }
    TechDetailResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/TechDetailVO' } }
    MomentDetailResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/MomentDetailVO' } }
...